Move the screen factory to the experience view controller
